requirements.txt
fastapi
uvicorn[standard]
python-docx
pydantic
python-multipart
reportlab
python-pptx
boto3
pymongo
python-dotenv
